Renaming package interpretations->fixedinterpretations and interface InterpretableLiteral->FixedInterpretationLiteral.
- In Predicate renaming attribute rank to arity consistently.
