+++ src/main/java/at/ac/tuwien/kr/alpha/common/fixedinterpretations/BindingPredicateInterpretation.java
@@ -1,4 +1,4 @@
-package at.ac.tuwien.kr.alpha.common.interpretations;
+package at.ac.tuwien.kr.alpha.common.fixedinterpretations;
 
 /**
  * This interface is used to mark predicate interpretations that might generate
