+++ src/main/java/at/ac/tuwien/kr/alpha/grounder/parser/ProgramParser.java
@@ -4,7 +4,7 @@ import at.ac.tuwien.kr.alpha.CustomErrorListener;
 import at.ac.tuwien.kr.alpha.antlr.ASPCore2Lexer;
 import at.ac.tuwien.kr.alpha.antlr.ASPCore2Parser;
 import at.ac.tuwien.kr.alpha.common.Program;
-import at.ac.tuwien.kr.alpha.common.interpretations.PredicateInterpretation;
+import at.ac.tuwien.kr.alpha.common.fixedinterpretations.PredicateInterpretation;
 import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.atn.PredictionMode;
 import org.antlr.v4.runtime.misc.ParseCancellationException;
